+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added function `util.format_duration()`
 - Added method `util.Cache.clear()`
+- Added method `cogs.voice.MediaItem.embed()`
+- Added property `cogs.voice.MediaItem.duration_str`
 
 ### Changed